# jarless/api/openapi.yaml
openapi: 3.0.2

info:
  version: 0.0.1
  title: Jarless API

servers:
  - url: http://localhost:5000
    description: Local

paths:

  /version:
    get:
      operationId: jarless.api.api_version
      summary: Retorna a versão da API
      responses:
        200:
          description: Success

  /executions/{package}:
    parameters:
      - name: package
        in: path
        description: Package
        required: true
        schema:
          type: string
    post:
      operationId: jarless.api.executions.create_execution
      summary: Create a new execution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                inputs:
                  additionalProperties: true
                  nullable: true
      responses:
        201:
          description: Created

  /executions/{package}/{task_id}/add/output/value:
    parameters:
      - name: package
        in: path
        description: Package
        required: true
        schema:
          type: string
      - name: task_id
        in: path
        description: Task ID
        required: true
        schema:
          type: string
      - name: secrets
        in: query
        description: Permission
        schema:
          type: string

    post:
      operationId: jarless.api.executions.add_output_value
      summary: Add an output for a given task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: myoutput
                value:
                  oneOf:
                    - type: string
                    - type: number
                    - type: boolean
                    - type: array
                    - type: object
                      example: 100
      responses:
        201:
          description: Added

  /executions/{package}/{task_id}/add/output/file:
    parameters:
      - name: package
        in: path
        description: Package
        required: true
        schema:
          type: string
      - name: task_id
        in: path
        description: Task ID
        required: true
        schema:
          type: string
      - name: secrets
        in: query
        description: Permission
        schema:
          type: string

    post:
      operationId: jarless.api.executions.add_output_file
      summary: Add an output for a given task
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: output1
                filedata:
                  type: string
                  format: binary

      responses:
        201:
          description: Added

  # /upload/{package}:
  #   parameters:
  #     - name: package
  #       in: path
  #       description: Package
  #       required: true
  #       schema:
  #         type: string

  #   post:
  #     operationId: jarless.api.executions.add_output_file
  #     summary: Add an output for a given task
  #     requestBody:
  #       content:
  #         multipart/form-data:
  #           schema:
  #             type: object
  #             properties:
  #               filedata:
  #                 type: string
  #                 format: binary

  #     responses:
  #       201:
  #         description: Added